+++ packages/cdn-icon/src/Component.tsx
@@ -24,7 +24,7 @@ export const CDNIcon: React.FC<CDNIconProps> = ({ name, color, dataTestId }) => 
         const xhr = new XMLHttpRequest();
         xhr.open('GET', `https://alfabank.st/icons/${name}.svg`);
         xhr.send();
-        xhr.onload = function() {
+        xhr.onload = function onload() {
             const svg = xhr.response;
             if (svg.startsWith('<svg')) setIcon(svg);
         };